+++ tests/helpers/test_log.py
@@ -212,7 +212,7 @@ class TestLog:
             ),
         ),
     )
-    # pylint: disable=too-many-arguments
+    # pylint: disable=too-many-arguments, too-many-positional-arguments
     def test_configure_logging(
         self,
         config: Config,